requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
azure-ai-vision-imageanalysis==1.0.0b1
azure-cognitiveservices-speech==1.34.0
openai==1.12.0
jinja2==3.1.3